Normalize data to avoid overflows
Dates converted with date2num() are 6-digit numbers,
of course, if they are used as x in an exponential
function - they produce an overflow.

Numbers were subtracted with the smallest date and
then divided with largest, so we'll have a range of
0 < x < 1.

We can revert this normalization to plot fitted
data on current data.