Create a polynomial fit
For the purpose of checking how it fits against the data
and how compares to exponential fit.

I commented lines for the plotting as poly fit does not fit correctly, but maybe
they could be used later. Maybe it can, with changing parameters and order, but
exponential function fits from the get-go and I will leave it.